+++ illico/asymptotic_wilcoxon.py
@@ -1,4 +1,3 @@
-import math
 from typing import Literal
 
 import anndata as ad
@@ -6,11 +5,13 @@ import numpy as np
 import pandas as pd
 from joblib import Parallel, delayed
 from loguru import logger
+from numba import set_num_threads
 from scipy import sparse
 from tqdm.auto import tqdm
 
 from illico.utils.compile import _precompile
 from illico.utils.groups import GroupContainer, encode_and_count_groups
+from illico.utils.math import compute_batch_bounds
 from illico.utils.memory import log_memory_usage
 from illico.utils.ranking import check_indices_sorted_per_parcel
 from illico.utils.registry import (
@@ -20,8 +21,7 @@ from illico.utils.registry import (
     nb_dispatcher_registry,
     rs_dispatcher_registry,
 )
-
-# from illico.utils.math import _warn_log1p
+from illico.utils.scanpy import format_illico_results_for_scanpy
 
 __all__ = ["asymptotic_wilcoxon"]
 
@@ -36,6 +36,7 @@ def operator(
     use_continuity: bool,
     alternative: str,
     tie_correct: bool,
+    exp_post_agg: bool,
     use_rust: bool,
     results: np.ndarray,
 ):
@@ -61,13 +62,14 @@ def operator(
     # Convert to numba-compatible format
     X = data_handler.to_nb(fetched_data)
     # Call the dispatcher
-    pvalues, statistics, fold_change = dispatcher(
+    pvalues, statistics, zscores, fold_change = dispatcher(
         X,
         *bounds,
         group_container,
         is_log1p,
         use_continuity,
         tie_correct,
+        exp_post_agg,
         alternative,
     )
     # Copy results into the shared array, do it thread-wise for cleaner garbage collection and speed
@@ -76,7 +78,8 @@ def operator(
     # The copy should be GIL-free
     results[:, lb:ub, 0] = pvalues
     results[:, lb:ub, 1] = statistics
-    results[:, lb:ub, 2] = fold_change  # Technically Rust returns f32, but numpy handles casting here
+    results[:, lb:ub, 2] = zscores
+    results[:, lb:ub, 3] = fold_change  # Technically Rust returns f32, but numpy handles casting here
     return (lb, ub)
 
 
@@ -90,10 +93,13 @@ def asymptotic_wilcoxon(
     alternative: str = "two-sided",
     use_continuity: bool = True,
     tie_correct: bool = True,
+    exp_post_agg: bool = False,
     layer: str | None = None,
     precompile: bool = True,
     use_rust: bool = True,
-):
+    return_as_scanpy: bool = False,
+    corr_method: Literal["benjamini-hochberg", "bonferroni"] = "benjamini-hochberg",
+) -> pd.DataFrame | dict:
     """Perform asymptotic Mann-Whitney tests for differential gene expression.
 
     Mann-Whitney test is the same as Wilcoxon rank-sum test.
@@ -125,20 +131,38 @@ def asymptotic_wilcoxon(
         Whether to apply continuity correction.
     tie_correct : bool, default=True
         Whether to apply tie correction in the test statistic.
+    exp_post_agg : bool, default=False
+        Whether to exponentiate the fold change after aggregation. This is relevant if the input data is log1p. See documentation for details.
+        Note that `scanpy.rank_genes_groups` assumes the data to be log1p, and exponentiates post aggregation by default.
     layer : str or None, default=None
         Layer in `adata.layers` to use for the data. If `None`, uses `adata.X`.
     precompile : bool, default=True
         Whether to precompile necessary functions for performance. It is recommended to set this to `True`.
     use_rust : bool, default=True
         Whether to use the Rust implementation of the test. If `False`, uses the Numba implementation.
+    return_as_scanpy : bool, default=False
+        Whether to return results in a format compatible with Scanpy's `rank_genes_groups` function.
+        If yes, the output is a dictionary that can be attached to the `adata` object like this:
+        `adata.uns['rank_genes_groups'] = asymptotic_wilcoxon(..., return_as_scanpy=True)`
+    corr_method: str, default="benjamini-hochberg"
+        Method to use for multiple testing correction. One of 'benjamini-hochberg' or 'bonferroni'.
+
 
     Returns
     -------
-    pd.DataFrame
+    Either one of pd.DataFrame or Dict, depending on the value of `return_as_scanpy`:
         A DataFrame with MultiIndex (pert, feature) containing columns:
         - 'p_value': P-value from the Mann-Whitney test
         - 'statistic': Test statistic (U-statistic)
+        - 'z-scores': Test z-score
         - 'fold_change': Fold change between groups
+        Or a dictionary formatted for Scanpy's `rank_genes_groups` results, containing:
+        - 'params': Dictionary of parameters used for the test
+        - 'names': Record array of gene names sorted by significance for each group
+        - 'scores': Record array of test statistics sorted by significance for each group
+        - 'pvals': Record array of p-values sorted by significance for each group
+        - 'pvals_adj': Record array of adjusted p-values sorted by significance for each group
+        - 'logfoldchanges': Record array of log2 fold changes sorted by significance for each group
 
     Raises
     ------
@@ -200,6 +224,7 @@ def asymptotic_wilcoxon(
 
     # Check that the input CSR is sorted.
     if isinstance(X, sparse.csr_matrix):
+        set_num_threads(n_threads)  # Set the number of threads for Numba to use in the check function
         if not check_indices_sorted_per_parcel(X.indices, X.indptr):
             raise ValueError(
                 "Input data matrix indices are not sorted. This is very unusual and may lead to incorrect results. "
@@ -227,25 +252,11 @@ def asymptotic_wilcoxon(
     # Allocate the results dataframes
     cols = pd.Series(adata.var_names, name="feature", dtype=str)
     rows = pd.Series(unique_raw_groups, name="pert", dtype=str)
-    results = np.empty((len(rows), len(cols), 3), dtype=np.float64)
+    results = np.empty((len(rows), len(cols), 4), dtype=np.float64)
 
-    # Adapt batch size to leverage multithreading regarding the number of genes, if requested
-    if n_genes < 256:
-        batch_size = n_genes  # No batching for small number of genes
-        n_threads = 1  # No multithreading for small number of genes
-        iterator = [[0, n_genes]]
-    elif isinstance(batch_size, int):
-        batch_size = min(batch_size, math.ceil(n_genes / n_threads))
-        bounds = np.append(np.arange(0, n_genes, batch_size), n_genes)
-        iterator = list(zip(bounds[:-1], bounds[1:]))
-    elif batch_size == "auto":
-        n_dispatches = max(int(n_genes / 256 / n_threads), 1)  # Aim for approximately 256 genes per chunk
-        splits = np.array_split(np.arange(n_genes + 1), indices_or_sections=n_threads * n_dispatches)
-        iterator = [split[[0, -1]] for split in splits]
-        batch_size = int(np.ceil(n_genes / (n_dispatches * n_threads)))
-    else:
-        raise ValueError(f"Invalid batch_size value: {batch_size}. Must be 'auto' or an integer.")
-    logger.trace(f"Using batch size of {batch_size} for {n_threads} threads and {n_genes} genes.")
+    # Compute the batch bounds for each thread
+    iterator, batch_size = compute_batch_bounds(n_genes, batch_size, n_threads)
+    logger.trace(f"Processing {n_genes} genes through {len(iterator)} batches with {n_threads} threads.")
 
     # Compute estimated mem footprint
     _ = log_memory_usage(data_handler, group_container, batch_size, n_threads)
@@ -265,6 +276,7 @@ def asymptotic_wilcoxon(
                     use_continuity,
                     alternative,
                     tie_correct,
+                    exp_post_agg,
                     use_rust,
                     results,
                 )
@@ -272,11 +284,22 @@ def asymptotic_wilcoxon(
             ):
                 pbar.update(group_container.counts.size * (ub - lb))
 
+    if not return_as_scanpy:
         # Return a pd.DataFrame to index results
         results = pd.DataFrame(
-            data=results.reshape(-1, 3),
+            data=results.reshape(-1, 4),
             index=pd.MultiIndex.from_product([rows, cols], names=["pert", "feature"]),
-            columns=["p_value", "statistic", "fold_change"],
+            columns=["p_value", "statistic", "z_score", "fold_change"],
+        )
+    else:
+        # Return a dict formatted for Scanpy's rank_genes_groups results
+        results = format_illico_results_for_scanpy(
+            adata=adata,
+            reference=reference,
+            group_keys=group_keys,
+            layer=layer,
+            values=results,
+            corr_method=corr_method,
         )
 
     return results