+++ src/sparse_ovo.rs
@@ -19,6 +19,7 @@ pub fn single_group_sparse_ovo_mwu_kernel<D: SparseFloat, I: SparseIndex>(
     alternative: &String,
     mut p_values: ArrayViewMut1<f64>,
     mut u_stats: ArrayViewMut1<f64>,
+    mut zscores: ArrayViewMut1<f64>,
 ) -> Result<(), String> {
     let n_cols_ctrl = ctrl.shape.1;
     let n_ctrl = ctrl.shape.0 as f64;
@@ -43,10 +44,8 @@ pub fn single_group_sparse_ovo_mwu_kernel<D: SparseFloat, I: SparseIndex>(
         n_zeros_tgt[j] = n_tgt - (tgt.indptr[j + 1] - tgt.indptr[j]).to_usize() as f64;
     }
 
-    // Pre allocate results
     let mu = n_ctrl * n_tgt / 2.;
-
-    let u_base = (n_ctrl * n_tgt) + n_tgt * (n_tgt + 1.) / 2.;
+    let remainder = n_tgt * (n_tgt + 1.) / 2.;
     for j in 0..n_cols_ctrl {
         let n_zeros_total = n_zeros_ctrl[j] + n_zeros_tgt[j];
         let (lbc, ubc) = (ctrl.indptr[j].to_usize(), ctrl.indptr[j + 1].to_usize());
@@ -66,9 +65,9 @@ pub fn single_group_sparse_ovo_mwu_kernel<D: SparseFloat, I: SparseIndex>(
         tiesum += n_zeros_total.powi(3) - n_zeros_total;
 
         // Compute U-stats
-        let u = u_base - ranksum;
+        let u = ranksum - remainder;
 
-        let pv = compute_pvalue(
+        let (pv, z) = compute_pvalue(
             n_ctrl,
             n_tgt,
             n_total,
@@ -80,6 +79,7 @@ pub fn single_group_sparse_ovo_mwu_kernel<D: SparseFloat, I: SparseIndex>(
         )?;
         p_values[j] = pv;
         u_stats[j] = u;
+        zscores[j] = z;
     }
 
     Ok(())
@@ -91,7 +91,7 @@ pub fn multigroup_sparse_ovo_mwu_kernel<D: SparseFloat, I: SparseIndex>(
     use_continuity: bool,
     tie_correct: bool,
     alternative: String,
-) -> Result<(Array2<f64>, Array2<f64>), String> {
+) -> Result<(Array2<f64>, Array2<f64>, Array2<f64>), String> {
     if grpc.encoded_ref_group < 0 {
         return Err(format!(
             "Encoded ref group can not be negative. Received {}.",
@@ -109,31 +109,35 @@ pub fn multigroup_sparse_ovo_mwu_kernel<D: SparseFloat, I: SparseIndex>(
     let n_groups = grpc.counts.len();
     let mut pvalues = Array2::zeros((n_groups, x.shape.1));
     let mut u_stats = Array2::zeros((n_groups, x.shape.1));
+    let mut zscores = Array2::zeros((n_groups, x.shape.1));
     for group_idx in 0..n_groups {
         if group_idx == encoded_ref_group {
             pvalues.row_mut(group_idx).fill(1.);
             u_stats.row_mut(group_idx).fill(-1.);
-        }
-        // Chunk the target
-        let start = grpc.indptr[group_idx as usize];
-        let end = grpc.indptr[group_idx as usize + 1];
-        let tgt_indices = grpc.indices.slice(s![start..end]);
-        let mut tgt_chunk = x.index_rows_into_csc(tgt_indices)?;
-        tgt_chunk.sort_columns_inplace()?;
+            zscores.row_mut(group_idx).fill(0.);
+        } else {
+            // Chunk the target
+            let start = grpc.indptr[group_idx as usize];
+            let end = grpc.indptr[group_idx as usize + 1];
+            let tgt_indices = grpc.indices.slice(s![start..end]);
+            let mut tgt_chunk = x.index_rows_into_csc(tgt_indices)?;
+            tgt_chunk.sort_columns_inplace()?;
 
-        // Now compute p-values and u-stats
-        _ = single_group_sparse_ovo_mwu_kernel(
-            &control_chunk,
-            tgt_chunk,
-            use_continuity,
-            tie_correct,
-            &alternative,
-            pvalues.row_mut(group_idx),
-            u_stats.row_mut(group_idx),
-        )?;
+            // Now compute p-values and u-stats
+            _ = single_group_sparse_ovo_mwu_kernel(
+                &control_chunk,
+                tgt_chunk,
+                use_continuity,
+                tie_correct,
+                &alternative,
+                pvalues.row_mut(group_idx),
+                u_stats.row_mut(group_idx),
+                zscores.row_mut(group_idx),
+            )?;
+        }
     }
 
-    Ok((pvalues, u_stats))
+    Ok((pvalues, u_stats, zscores))
 }
 
 pub fn csc_ovo_mwu_kernel_over_contiguous_col_chunk<'py, D: SparseFloat, I: SparseIndex>(
@@ -144,16 +148,17 @@ pub fn csc_ovo_mwu_kernel_over_contiguous_col_chunk<'py, D: SparseFloat, I: Spar
     is_log1p: bool,
     use_continuity: bool,
     tie_correct: bool,
+    exp_post_agg: bool,
     alternative: String,
-) -> Result<(Array2<f64>, Array2<f64>, Array2<f32>), String> {
+) -> Result<(Array2<f64>, Array2<f64>, Array2<f64>, Array2<f32>), String> {
     let chunk = x.contig_cols_into_csr(chunk_lb, chunk_ub)?;
 
-    let (pvalues, u_stats) =
+    let (pvalues, u_stats, zscores) =
         multigroup_sparse_ovo_mwu_kernel(&chunk, &grpc, use_continuity, tie_correct, alternative)?;
 
-    let fc = csr_fold_change(&chunk, &grpc, is_log1p)?;
+    let fc = csr_fold_change(&chunk, &grpc, is_log1p, exp_post_agg)?;
 
-    Ok((pvalues, u_stats, fc))
+    Ok((pvalues, u_stats, zscores, fc))
 }
 
 pub fn csr_ovo_mwu_kernel_over_contiguous_col_chunk<'py, D: SparseFloat, I: SparseIndex>(
@@ -164,105 +169,29 @@ pub fn csr_ovo_mwu_kernel_over_contiguous_col_chunk<'py, D: SparseFloat, I: Spar
     is_log1p: bool,
     use_continuity: bool,
     tie_correct: bool,
+    exp_post_agg: bool,
     alternative: String,
-) -> Result<(Array2<f64>, Array2<f64>, Array2<f32>), String> {
+) -> Result<(Array2<f64>, Array2<f64>, Array2<f64>, Array2<f32>), String> {
     let chunk = x.contig_cols_into_csr(chunk_lb, chunk_ub)?;
 
-    let (pvalues, u_stats) =
+    let (pvalues, u_stats, zscores) =
         multigroup_sparse_ovo_mwu_kernel(&chunk, &grpc, use_continuity, tie_correct, alternative)?;
 
-    let fc = csr_fold_change(&chunk, &grpc, is_log1p)?;
+    let fc = csr_fold_change(&chunk, &grpc, is_log1p, exp_post_agg)?;
 
-    Ok((pvalues, u_stats, fc))
+    Ok((pvalues, u_stats, zscores, fc))
 }
-
-// macro_rules! ovo_mwu_kernel_over_contiguous_col_chunk {
-//     (x:expr, $grpc:expr, $chunk_lb:expr, $chunk_ub:expr, $is_log1p:expr, $use_continuity:expr, $tie_correct:expr, $alternative: expr) => {
-//         let chunk = x.contig_cols_into_csr($chunk_lb, $chunk_ub);
-//         let (pvalues, u_stats) =
-//             multigroup_sparse_ovo_mwu_kernel(&chunk, &grpc, use_continuity, tie_correct, alternative)?;
-
-//         let fc = csr_fold_change(&chunk, &grpc, is_log1p)?;
-
-//         (pvalues, u_stats, fc)
-//     };
-// }
 
 type PyArr2f32<'py> = Bound<'py, PyArray2<f32>>;
 type PyArr2f64<'py> = Bound<'py, PyArray2<f64>>;
-
-// #[pyfunction]
-// pub fn csc_ovo_mwu_kernel_over_contiguous_col_chunk_rust<'py>(
-//     py: Python<'py>,
-//     x: PyCSCMatrix<'py>,
-//     chunk_lb: usize,
-//     chunk_ub: usize,
-//     grpc: GroupContainerNamedTuple,
-//     is_log1p: bool,
-//     use_continuity: bool,
-//     tie_correct: bool,
-//     alternative: String,
-// ) -> PyResult<(PyArr2f64<'py>, PyArr2f64<'py>, PyArr2f32<'py>)> {
-//     let x = x.as_csc_matrix();
-//     let grpc = grpc.as_group_container();
-
-//     let (pvalues, u_stats, fc) = py
-//         .detach(|| {
-//             csc_ovo_mwu_kernel_over_contiguous_col_chunk(
-//                 &x,
-//                 grpc,
-//                 chunk_lb,
-//                 chunk_ub,
-//                 is_log1p,
-//                 use_continuity,
-//                 tie_correct,
-//                 alternative,
-//             )
-//         })
-//         .map_err(PyValueError::new_err)?;
-
-//     Ok((
-//         PyArray2::from_array(py, &pvalues),
-//         PyArray2::from_array(py, &u_stats),
-//         PyArray2::from_array(py, &fc),
-//     ))
-// }
-
-// macro_rules! run_csr_branch {
-//         ($x:expr, $py:expr, $grpc:expr, $chunk_lb:expr, $chunk_ub:expr, $is_log1p:expr, $use_continuity:expr, $tie_correct:expr, $alternative:expr, $dt:ty, $it:ty) => {{
-//                 let data = $x.data.extract::<PyReadonlyArray1<'py, $dt>>()?;
-//                 let indices = $x.indices.extract::<PyReadonlyArray1<'py, $it>>()?;
-//                 let indptr = $x.indptr.extract::<PyReadonlyArray1<'py, $it>>()?;
-
-//                 let csr = CSRMatrix {
-//                         data: data.as_array(),
-//                         indices: indices.as_array(),
-//                         indptr: indptr.as_array(),
-//                         shape: $x.shape,
-//                     };
-
-//                     $py.detach(|| {
-//                             csr_ovo_mwu_kernel_over_contiguous_col_chunk(
-//                                     &csr,
-//                                     $grpc,
-//                                     $chunk_lb,
-//                                     $chunk_ub,
-//                                     $is_log1p,
-//                                     $use_continuity,
-//                                     $tie_correct,
-//                                     $alternative,
-//                                 )
-//                             })
-//                             .map_err(PyValueError::new_err)
-//                         }};
-//     }
 
 // The extraction into PyArray + conversion to Array + compute has to be done in one single function, because dtypes are not known at compile time and pyfunctions dont accept generic traits.
 // Hence, it is not possible to have let's say a function returning a dtyped object: even PyAny.extract -> PyArray because PyArray has to be typed.
 // Previous implementation was 1/ FromPyObject's .extract returning a PyArray, 2/ then .as_csr returning an Array. None of those can be compiled in the dtype-agnostic setup.
 // Hence, conversion into pyarray, then conversion into arrays must happen in the same scope when dtype is known.
+#[rustfmt::skip]
 macro_rules! run_branch {
-    ($format:expr, $x:expr, $py:expr, $grpc:expr, $chunk_lb:expr, $chunk_ub:expr, $is_log1p:expr, $use_continuity:expr, $tie_correct:expr, $alternative:expr, $dt:ty, $it:ty) => {{
+    ($format:expr, $x:expr, $py:expr, $grpc:expr, $chunk_lb:expr, $chunk_ub:expr, $is_log1p:expr, $use_continuity:expr, $tie_correct:expr, $exp_post_agg:expr, $alternative:expr, $dt:ty, $it:ty) => {{
         let data = $x.data.extract::<PyReadonlyArray1<'py, $dt>>()?;
         let indices = $x.indices.extract::<PyReadonlyArray1<'py, $it>>()?;
         let indptr = $x.indptr.extract::<PyReadonlyArray1<'py, $it>>()?;
@@ -279,16 +208,8 @@ macro_rules! run_branch {
                 };
 
                 $py.detach(|| {
-                    // ovo_mwu_kernel_over_contiguous_col_chunk!(
                     csr_ovo_mwu_kernel_over_contiguous_col_chunk(
-                        &csr,
-                        $grpc,
-                        $chunk_lb,
-                        $chunk_ub,
-                        $is_log1p,
-                        $use_continuity,
-                        $tie_correct,
-                        $alternative,
+                        &csr, $grpc, $chunk_lb, $chunk_ub, $is_log1p, $use_continuity, $tie_correct, $exp_post_agg, $alternative,
                     )
                 })
                 .map_err(PyValueError::new_err)
@@ -303,14 +224,7 @@ macro_rules! run_branch {
 
                 $py.detach(|| {
                     csc_ovo_mwu_kernel_over_contiguous_col_chunk(
-                        &csc,
-                        $grpc,
-                        $chunk_lb,
-                        $chunk_ub,
-                        $is_log1p,
-                        $use_continuity,
-                        $tie_correct,
-                        $alternative,
+                        &csc, $grpc, $chunk_lb, $chunk_ub, $is_log1p, $use_continuity, $tie_correct, $exp_post_agg, $alternative,
                     )
                 })
                 .map_err(PyValueError::new_err)
@@ -320,6 +234,7 @@ macro_rules! run_branch {
     }};
 }
 
+#[rustfmt::skip]
 #[pyfunction]
 pub fn csr_ovo_mwu_kernel_over_contiguous_col_chunk_rust<'py>(
     py: Python<'py>,
@@ -330,68 +245,30 @@ pub fn csr_ovo_mwu_kernel_over_contiguous_col_chunk_rust<'py>(
     is_log1p: bool,
     use_continuity: bool,
     tie_correct: bool,
+    exp_post_agg: bool,
     alternative: String,
-) -> PyResult<(PyArr2f64<'py>, PyArr2f64<'py>, PyArr2f32<'py>)> {
+) -> PyResult<(
+    PyArr2f64<'py>,
+    PyArr2f64<'py>,
+    PyArr2f64<'py>,
+    PyArr2f32<'py>,
+)> {
     let grpc = grpc.as_group_container();
 
     let data_dtype: String = x.data.getattr("dtype")?.getattr("str")?.extract()?;
     let idx_dtype: String = x.indices.getattr("dtype")?.getattr("str")?.extract()?;
-    let (pv, u, fc) = match (data_dtype.as_str(), idx_dtype.as_str()) {
+    let (pv, u, z, fc) = match (data_dtype.as_str(), idx_dtype.as_str()) {
         ("f32" | "<f4", "i32" | "<i4") => run_branch!(
-            "CSR",
-            x,
-            py,
-            grpc,
-            chunk_lb,
-            chunk_ub,
-            is_log1p,
-            use_continuity,
-            tie_correct,
-            alternative,
-            f32,
-            i32
+            "CSR", x, py, grpc, chunk_lb, chunk_ub, is_log1p, use_continuity, tie_correct, exp_post_agg, alternative, f32, i32
         ),
         ("f64" | "<f8", "i32" | "<i4") => run_branch!(
-            "CSR",
-            x,
-            py,
-            grpc,
-            chunk_lb,
-            chunk_ub,
-            is_log1p,
-            use_continuity,
-            tie_correct,
-            alternative,
-            f64,
-            i32
+            "CSR", x, py, grpc, chunk_lb, chunk_ub, is_log1p, use_continuity, tie_correct, exp_post_agg, alternative, f64, i32
         ),
         ("f32" | "<f4", "i64" | "<i8") => run_branch!(
-            "CSR",
-            x,
-            py,
-            grpc,
-            chunk_lb,
-            chunk_ub,
-            is_log1p,
-            use_continuity,
-            tie_correct,
-            alternative,
-            f32,
-            i64
+            "CSR", x, py, grpc, chunk_lb, chunk_ub, is_log1p, use_continuity, tie_correct, exp_post_agg, alternative, f32, i64
         ),
         ("f64" | "<f8", "i64" | "<i8") => run_branch!(
-            "CSR",
-            x,
-            py,
-            grpc,
-            chunk_lb,
-            chunk_ub,
-            is_log1p,
-            use_continuity,
-            tie_correct,
-            alternative,
-            f64,
-            i64
+            "CSR", x, py, grpc, chunk_lb, chunk_ub, is_log1p, use_continuity, tie_correct, exp_post_agg, alternative, f64, i64
         ),
         _ => Err(PyValueError::new_err(format!(
             "Error casting data (only f32 and f64 supported, received {}) and indices (only int32 and int64 supported, received {}).",
@@ -402,10 +279,12 @@ pub fn csr_ovo_mwu_kernel_over_contiguous_col_chunk_rust<'py>(
     return Ok((
         PyArray2::from_array(py, &pv),
         PyArray2::from_array(py, &u),
+        PyArray2::from_array(py, &z),
         PyArray2::from_array(py, &fc),
     ));
 }
 
+#[rustfmt::skip]
 #[pyfunction]
 pub fn csc_ovo_mwu_kernel_over_contiguous_col_chunk_rust<'py>(
     py: Python<'py>,
@@ -416,68 +295,30 @@ pub fn csc_ovo_mwu_kernel_over_contiguous_col_chunk_rust<'py>(
     is_log1p: bool,
     use_continuity: bool,
     tie_correct: bool,
+    exp_post_agg: bool,
     alternative: String,
-) -> PyResult<(PyArr2f64<'py>, PyArr2f64<'py>, PyArr2f32<'py>)> {
+) -> PyResult<(
+    PyArr2f64<'py>,
+    PyArr2f64<'py>,
+    PyArr2f64<'py>,
+    PyArr2f32<'py>,
+)> {
     let grpc = grpc.as_group_container();
 
     let data_dtype: String = x.data.getattr("dtype")?.getattr("str")?.extract()?;
     let idx_dtype: String = x.indices.getattr("dtype")?.getattr("str")?.extract()?;
-    let (pv, u, fc) = match (data_dtype.as_str(), idx_dtype.as_str()) {
+    let (pv, u, z, fc) = match (data_dtype.as_str(), idx_dtype.as_str()) {
         ("f32" | "<f4", "i32" | "<i4") => run_branch!(
-            "CSC",
-            x,
-            py,
-            grpc,
-            chunk_lb,
-            chunk_ub,
-            is_log1p,
-            use_continuity,
-            tie_correct,
-            alternative,
-            f32,
-            i32
+            "CSC", x, py, grpc, chunk_lb, chunk_ub, is_log1p, use_continuity, tie_correct, exp_post_agg, alternative, f32, i32
         ),
         ("f64" | "<f8", "i32" | "<i4") => run_branch!(
-            "CSC",
-            x,
-            py,
-            grpc,
-            chunk_lb,
-            chunk_ub,
-            is_log1p,
-            use_continuity,
-            tie_correct,
-            alternative,
-            f64,
-            i32
+            "CSC", x, py, grpc, chunk_lb, chunk_ub, is_log1p, use_continuity, tie_correct, exp_post_agg, alternative, f64, i32
         ),
         ("f32" | "<f4", "i64" | "<i8") => run_branch!(
-            "CSC",
-            x,
-            py,
-            grpc,
-            chunk_lb,
-            chunk_ub,
-            is_log1p,
-            use_continuity,
-            tie_correct,
-            alternative,
-            f32,
-            i64
+            "CSC", x, py, grpc, chunk_lb, chunk_ub, is_log1p, use_continuity, tie_correct, exp_post_agg, alternative, f32, i64
         ),
         ("f64" | "<f8", "i64" | "<i8") => run_branch!(
-            "CSC",
-            x,
-            py,
-            grpc,
-            chunk_lb,
-            chunk_ub,
-            is_log1p,
-            use_continuity,
-            tie_correct,
-            alternative,
-            f64,
-            i64
+            "CSC", x, py, grpc, chunk_lb, chunk_ub, is_log1p, use_continuity, tie_correct, exp_post_agg, alternative, f64, i64
         ),
         _ => Err(PyValueError::new_err(format!(
             "Error casting data (only f32 and f64 supported, received {}) and indices (only int32 and int64 supported, received {}).",
@@ -488,64 +329,7 @@ pub fn csc_ovo_mwu_kernel_over_contiguous_col_chunk_rust<'py>(
     return Ok((
         PyArray2::from_array(py, &pv),
         PyArray2::from_array(py, &u),
+        PyArray2::from_array(py, &z),
         PyArray2::from_array(py, &fc),
     ));
 }
-// #[pyfunction]
-// pub fn csr_ovo_mwu_kernel_over_contiguous_col_chunk_rust<'py>(
-//     py: Python<'py>,
-//     x: PyCSRMatrix2<'py>,
-//     chunk_lb: usize,
-//     chunk_ub: usize,
-//     grpc: GroupContainerNamedTuple,
-//     is_log1p: bool,
-//     use_continuity: bool,
-//     tie_correct: bool,
-//     alternative: String,
-// ) -> PyResult<(PyArr2f64<'py>, PyArr2f64<'py>, PyArr2f32<'py>)> {
-//     let grpc = grpc.as_group_container();
-
-//     let data_dtype: String = x.data.getattr("dtype")?.getattr("str")?.extract()?;
-//     let idx_dtype: String = x.indices.getattr("dtype")?.getattr("str")?.extract()?;
-//     let (pv, u, fc) = match (data_dtype.as_str(), idx_dtype.as_str()) {
-//         ("f32" | "<f4", "i32" | "<i4") => {
-//             let data = x.data.extract::<PyReadonlyArray1<'py, f32>>()?;
-//             let indices = x.indices.extract::<PyReadonlyArray1<'py, i32>>()?;
-//             let indptr = x.indptr.extract::<PyReadonlyArray1<'py, i32>>()?;
-//             let csr = CSRMatrix { data: data.as_array(), indices: indices.as_array(), indptr: indptr.as_array(), shape: x.shape};
-//             let (pv, u, fc) = py.detach(|| {csr_ovo_mwu_kernel_over_contiguous_col_chunk(&csr, grpc, chunk_lb, chunk_ub, is_log1p, use_continuity, tie_correct, alternative)}).map_err(PyValueError::new_err)?;
-//             Ok((pv, u, fc))
-//         },
-//         ("f64" | "<f8", "i32" | "<i4") => {
-//             let data = x.data.extract::<PyReadonlyArray1<'py, f64>>()?;
-//             let indices = x.indices.extract::<PyReadonlyArray1<'py, i32>>()?;
-//             let indptr = x.indptr.extract::<PyReadonlyArray1<'py, i32>>()?;
-//             let csr = CSRMatrix { data: data.as_array(), indices: indices.as_array(), indptr: indptr.as_array(), shape: x.shape};
-//             let (pv, u, fc) = py.detach(|| {csr_ovo_mwu_kernel_over_contiguous_col_chunk(&csr, grpc, chunk_lb, chunk_ub, is_log1p, use_continuity, tie_correct, alternative)}).map_err(PyValueError::new_err)?;
-//             Ok((pv, u, fc))
-//         },
-//         ("f32" | "<f4", "i64" | "<i8") => {
-//             let data = x.data.extract::<PyReadonlyArray1<'py, f32>>()?;
-//             let indices = x.indices.extract::<PyReadonlyArray1<'py, i64>>()?;
-//             let indptr = x.indptr.extract::<PyReadonlyArray1<'py, i64>>()?;
-//             let csr = CSRMatrix { data: data.as_array(), indices: indices.as_array(), indptr: indptr.as_array(), shape: x.shape};
-//             let (pv, u, fc) = py.detach(|| {csr_ovo_mwu_kernel_over_contiguous_col_chunk(&csr, grpc, chunk_lb, chunk_ub, is_log1p, use_continuity, tie_correct, alternative)}).map_err(PyValueError::new_err)?;
-//             Ok((pv, u, fc))
-//         },
-//         ("f64" | "<f8", "i64" | "<i8") => {
-//             let data = x.data.extract::<PyReadonlyArray1<'py, f64>>()?;
-//             let indices = x.indices.extract::<PyReadonlyArray1<'py, i64>>()?;
-//             let indptr = x.indptr.extract::<PyReadonlyArray1<'py, i64>>()?;
-//             let csr = CSRMatrix { data: data.as_array(), indices: indices.as_array(), indptr: indptr.as_array(), shape: x.shape};
-//             let (pv, u, fc) = py.detach(|| {csr_ovo_mwu_kernel_over_contiguous_col_chunk(&csr, grpc, chunk_lb, chunk_ub, is_log1p, use_continuity, tie_correct, alternative)}).map_err(PyValueError::new_err)?;
-//             Ok((pv, u, fc))
-//         },
-//         _ => Err(PyValueError::new_err(format!("Error casting {} and {}", data_dtype, idx_dtype))),
-//     }?;
-//     return Ok((
-//         PyArray2::from_array(py, &pv),
-//         PyArray2::from_array(py, &u),
-//         PyArray2::from_array(py, &fc),
-//     ))
-
-// }