+++ src/stats.rs
@@ -10,32 +10,31 @@ pub fn compute_pvalue(
     mu: f64,
     contin_corr: f64,
     alternative: &String,
-) -> Result<f64, String> {
+) -> Result<(f64, f64), String> {
     let tie_corr: f64 = 1.0 - tie_sum / (n * (n - 1.) * (n + 1.));
     if tie_corr > 1e-9 {
         let sigma: f64 = (n_ref * n_tgt * (n_ref + n_tgt + 1.) / 12.0 * tie_corr).powf(0.5);
 
         match alternative.as_str() {
             "two-sided" => {
-                let min_u = U.min(n_ref * n_tgt - U);
-                let delta = min_u - mu;
-                let z = (delta.abs() + delta.signum() * contin_corr) / sigma;
-                return Ok(erfc(z / (2.0 as f64).sqrt()));
+                let delta = U - mu;
+                let z = (delta - delta.signum() * contin_corr) / sigma;
+                return Ok((erfc(z.abs() / (2.0 as f64).sqrt()), z));
             }
             "greater" => {
                 let delta = U - mu;
                 let z = (delta - contin_corr) / sigma;
-                return Ok(0.5 * erfc(z / (2.0 as f64).sqrt()));
+                return Ok((0.5 * erfc(z / (2.0 as f64).sqrt()), z));
             }
             "less" => {
                 let delta = U - mu;
                 let z = (delta + contin_corr) / sigma;
-                return Ok(0.5 * erfc(-z / (2.0 as f64).sqrt()));
+                return Ok((0.5 * erfc(-z / (2.0 as f64).sqrt()), z));
             }
             _ => Err(format!("Invalid alternative: received {alternative}.")),
         }
     } else {
-        return Ok(1.0 as f64);
+        return Ok((1.0, 0.));
     }
 }
 
@@ -49,8 +48,7 @@ pub fn compute_pvalue_rust(
     mu: f64,
     contin_corr: f64,
     alternative: String,
-) -> PyResult<f64> {
-    // let p_value: f64 = compute_pvalue(n_ref, n_tgt, n, tie_sum, U, mu, contin_corr, &alternative)?;
+) -> PyResult<(f64, f64)> {
     compute_pvalue(
         n_ref as f64,
         n_tgt as f64,