+++ changelog.md
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+Version 0.4.0
+------------
+- Added option to return scanpy-friendly output with `return_as_scanpy` arg. `asymptotic_wilcoxon` returns either:
+    - A `pandas.DataFrame` with columns `feature`, `p_value`, `fold_change`, and `statistic` (default), if `return_as_scanpy=False`
+    - A dictionary containing the same keys as `scanpy.tl.rank_genes_groups`, if `return_as_scanpy=True`. Similarly as scanpy, genes are ordered by decreasing z-score.
+- Improved the batching mechanism, fixed the 'auto' mode that was excluding the very last gene in previous versions.
+
 Version 0.3.0
 ------------
 - Rust backend is available for all tests. Compare Rust vs Numba with `poetry run pytest-benchmark compare 0003 0005`: